重构代码，修复拼写问题 refactor code; fix typo
