+++ DOF5RobotControl_GUI.csproj
@@ -17,9 +17,6 @@
     <None Update="libDOF5RobotControl.dll">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </None>
-    <None Update="TestDll.dll">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </None>
   </ItemGroup>
 
 </Project>
