--- JointPosition.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace DOF5RobotControl_GUI
-{
-    internal class JointPosition
-    {
-        public string ColorName { get; set; }
-    }
-}
